+++ spec/lib/glimmer/data_binding/observable_array_spec.rb
@@ -48,6 +48,30 @@ module Glimmer
           expect(@fired).to eq(false)
         end
         
+        it 'adds recursive observer to an array having array elements (without element properties) using alternate syntax' do
+          @fired = false
+          observer = Observer.proc {
+            @fired = true
+          }
+          array = [['a', 'b']]
+          array.singleton_class.include(ObservableArray)
+          observer.observe(array, recursive: true)
+          expect(@fired).to eq(false)
+          array[0][0] = 'c'
+          expect(@fired).to eq(true)
+          
+          @fired = false
+          old_array = array[0]
+          array[0] = ['c', 'd']
+          expect(@fired).to eq(true)
+          
+          @fired = false
+          old_array << 'e'
+          expect(@fired).to eq(false)
+          array[0] << 'e'
+          expect(@fired).to eq(false)
+        end
+        
         it 'adds non-recursive observer to an array having array elements (without element properties)' do
           @fired = false
           observer = Observer.proc {